organize new dashboard componenets - add status-buttons component
