+++ src/components/dashboard/event-view/recent-card/event-button-modal/warn-remove-modal/WarnRemoveModal.test.js
