+++ src/components/dashboard/event-view/recent-card/recent-card.test.js
