+++ src/components/shared/grid-structure/header/header-partition-persistent/headerPartitionPersistent.jsx
@@ -35,7 +35,7 @@ function PersistentHeader(props) {
         <Icon icon={magnifyingGlass} />
       </Typography> */}
 
-            <AccountDrawer {...props} />
+            {/* <AccountDrawer {...props} /> */}
         </section>
     );
 }
